requirements.txt
aiofiles
Django
django-admin-sortable2
django-ckeditor
django-cleanup
django-dotenv
djangorestframework
drf-spectacular
drf-spectacular-sidecar
rich
uvicorn